requirements.txt
keras~=2.6.0
numpy~=1.21.5
six~=1.16.0
h5py~=3.1.0
PyYAML~=6.0
scipy~=1.7.3
tensorflow~=2.5.0
requests~=2.27.1
Pillow~=9.0.1
pytest~=7.1.2
pandas~=1.2.3
matplotlib~=3.5.1
scikit-learn~=0.23.2
setuptools~=58.0.4
hl7~=0.4.5
bcolors~=1.0.4
Flask~=2.3.2
scikit-image~=0.19.2
pydicom~=2.0.0
opencv-python~=4.7.0.72
imageio~=2.16.1
psutil~=5.9.1
python-dateutil~=2.8.2
tqdm~=4.64.1
pyod~=1.0.4
mahotas~=1.4.11
tornado~=6.1
pylibjpeg~=1.0.0
hyperopt~=0.2.7
statsmodels~=0.13.2